Split closure out of object.c/.h and veneer.c/.h
