+++ src/classes/CMakeLists.txt
@@ -1,10 +1,11 @@
 target_sources(morpho
     PRIVATE
-        array.c  array.h
-        cmplx.c  cmplx.h
-        file.c   file.h
-        list.c   list.h
-        range.c  range.h
-        strng.c  strng.h
-        system.c system.h
+        array.c    array.h
+        closure.c  closure.h
+        cmplx.c    cmplx.h
+        file.c     file.h
+        list.c     list.h
+        range.c    range.h
+        strng.c    strng.h
+        system.c   system.h
 )
